requirements.txt
absl-py==1.2.0
affine==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.13
alembic==1.8.1
aniso8601==9.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astroid==2.15.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
awscli==1.25.13
Babel==2.9.0
backcall==0.2.0
basemap==1.3.6
basemap-data==1.3.2
beautifulsoup4==4.11.1
black==23.3.0
bleach==5.0.1
boltons==23.0.0
boto3==1.24.59
botocore==1.27.59
Bottleneck==1.3.7
cachetools==5.2.0
certifi==2022.12.7
cf-xarray==0.8.0
cffi==1.15.1
cfgv==3.3.1
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.30.0
cloudpickle==2.2.1
colorama==0.4.5
colorclass==2.2.2
constants==0.6.0
coverage==7.2.2
crispy-bootstrap5==0.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.28
dask==2023.4.1
datasets==2.4.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.4
dj-database-url==2.0.0
dj-stripe==2.7.3
Django==4.2.3
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.6.0
django-coverage-plugin==3.0.0
django-crispy-forms==2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-storages==1.13.1
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.13.1
djangorestframework-stubs==1.10.0
docit==1.0.0.2.post3
docopt==0.6.2
docutils==0.19
dparse==0.6.2
drf-spectacular==0.26.1
entrypoints==0.4
env==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==0.8.3
factory-boy==3.2.1
Faker==18.4.0
fastjsonschema==2.16.1
filelock==3.7.1
flake8==6.0.0
flake8-isort==6.0.0
flasgger==0.9.5
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.34.4
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
gensim==4.2.0
gevent==22.10.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
greenlet==1.1.3.post0
grpcio==1.47.0
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.8.1
identify==2.5.22
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.7.1
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.13
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonfield==3.1.0
jsonpickle==3.0.1
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
livereload==2.6.3
llvmlite==0.40.0
lmoments3==1.0.5
locket==1.0.0
lxml==4.9.1
mailjet-rest==1.3.4
Mako==1.2.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
MetPy==1.4.1
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.13
mypy==1.1.1
mypy-extensions==1.0.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netCDF4==1.6.2
nltk==3.7
nodeenv==1.7.0
notebook==6.4.12
num2word==1.0.1
num2words==0.5.11
numba==0.57.0
numpy==1.23.1
oauthlib==3.2.0
openpyxl==3.0.10
packaging==23.1
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathspec==0.9.0
patsy==0.5.3
pbr==5.9.0
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.21
pip-upgrader==1.4.15
platformdirs==2.5.2
pluggy==1.0.0
pooch==1.7.0
pre-commit==3.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2==2.9.3
psycopg2-binary==2.9.5
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.4.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.18.1
pyshp==2.3.1
PySocks==1.7.1
pytest==7.2.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-sugar==0.9.6
python-dateutil==2.8.2
python-decouple==3.4
python-docx==0.8.11
python-dotenv==0.21.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
pywin32==304
pywinpty==2.0.6
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
rasterio==1.3.6
redis==4.5.4
regex==2022.7.25
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
responses==0.18.0
rioxarray==0.14.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
safety==2.1.1
scikit-learn==1.1.2
scipy==1.9.0
Send2Trash==1.8.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0
slippers==0.5.0
smart-open==6.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
snuggs==1.4.7
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spicy==0.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.4
stack-data==0.3.0
statsmodels==0.13.5
stevedore==4.0.0
stripe==4.2.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminado==0.15.0
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.4.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
types-requests==2.28.11.17
types-urllib3==1.26.25.10
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.2
Werkzeug==2.2.3
whitenoise==6.5.0
widgetsnbextension==3.6.1
wrapt==1.15.0
WTForms==3.0.1
xarray==2022.12.0
xclim==0.42.0
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.5.0
